backend/requirements.txt
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
PyJWT==2.8.1
djangorestframework-simplejwt==5.3.2
Pillow==10.1.0
gunicorn==21.2.0
drf-spectacular==0.26.5
django-environ==0.21.0
